+++ src/sentry/debug_files/objectstore_migration/__init__.py
@@ -1,0 +1,13 @@
+from sentry.debug_files.objectstore_migration.main import start_migration
+from sentry.debug_files.objectstore_migration.utils import (
+    ObjectstoreIntegrityError,
+    freeze_high_water_mark,
+    migrate_debug_file,
+)
+
+__all__ = (
+    "ObjectstoreIntegrityError",
+    "freeze_high_water_mark",
+    "migrate_debug_file",
+    "start_migration",
+)
